WIP change all verbose types to types.h types

+++ include/std/strings.h
@@ -3,7 +3,7 @@
 
 #include <data/types.h>
 
-unsigned int str_len(const char* str);
+size_t str_len(const char* str);
 char* strncpy(char* dest, const char* src, uint64 n);
 
 #endif /* STD_STRINGS_H */
